moved commandargs class to right folder
